requirements.txt
aniso8601==3.0.2
click==6.7
DateTime==4.2
Flask==1.0.2
Flask-HTTPAuth==3.2.4
Flask-PyMongo==2.1.0
Flask-RESTful==0.3.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
passlib==1.7.1
pymongo==3.7.1
python-dotenv==0.9.1
pytz==2018.5
six==1.11.0
Werkzeug==0.14.1
zope.interface==4.5.0
